Fix dupe-detection failure in FileMerger
The merger could fail to detect dupe of a new entry and commented-out
entry in the destination file when the new entry does not have
"__duplicate__" metadata from the initial dupe detection in the caller.

This can happen when the heuristic comparator in beangulp ignores
interpolated postings and thus it can fail to find a common posting
among an existing and the new entries even when the user has manually
commented the same duplicate entry in a past.

It is encouraged for the users to avoid units interpolation in such
potentially duplicating entries.